+++ src/test/java/tpi/backend/e_commerce/categoryTests/integrationTests/PruebaN07.java
@@ -1,5 +1,5 @@
 package tpi.backend.e_commerce.categoryTests.integrationTests;
 
-public class TestRecoveryCategorty {
+public class PruebaN07 {
     
 }
